Initial User Support
The plan is to implement Authentication Middleware in order
to authenticate incoming requests across the app.

In order to do that, we need a mechanism to allow users to signup.

Once we have this merged, I can use TSec to implement auth
middleware using JWT, and then authenticate all requests.  That
will be a future PR however.

This PR gives us what we need for the first step.