Refactor a few methods to use OptionT monad transformer
Also remove a couple unnecessary for comprehensions

Motiviation:
Note that `OptionT[F, A]` is a functor if `F` is a `Functor`, and it's a monad
if `F` is a monad. So entering the context of `OptionT[F, A]` for some A still
has the same semantics that we are used to dealing with in `F`, without the
need for nesting to get inside `Option` which we have for `F[Option[A]]`.

[See the docs here](https://typelevel.org/cats/datatypes/optiont.html)